requirements.txt
albumentations==0.4.3
clip==0.2.0
einops==0.8.0
h5py==3.10.0
kornia==0.7.2
natsort==8.4.0
nibabel==5.2.1
numpy==1.24.3
omegaconf==2.3.0
opencv_python==4.1.2.30
opencv_python_headless==4.8.1.78
packaging==24.0
Pillow==10.0.1
Pillow==10.3.0
pudb==2019.2
pytorch_lightning==1.4.2
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.9.1
SimpleITK==2.3.1
SimpleITK==2.3.1
taming_transformers==0.0.1
torch==2.0.1+cu118
torchio==0.19.3
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.6.0
